improve event end datetime display
This will introduce a bug around the way timezones behave around DST periods
The temporary solution subtracts 61 minutes from the UTC time. During DST this sets the time to 23:59:59 UTC and during GMT 22:59:59.
If a user books a desk with a start time after 22:59:59 during the winter then the booking will not show.